requirements.txt
flask==2.2.3
google-api-python-client==2.125.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
werkzeug==2.2.3
python-dotenv==1.0.0
GitPython==3.1.43